% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lm.R
\name{fit_lm}
\alias{fit_lm}
\title{Fit linear model, additionally report weights, additional metrics}
\usage{
fit_lm(variables, target, data, model_has_intercept = TRUE, concise = TRUE)
}
\arguments{
\item{variables}{vector of variable names}

\item{target}{target variable name}

\item{data}{data frame}

\item{model_has_intercept}{whether model has intercept}

\item{concise}{whether to return concise output}
}
\value{
list
}
\description{
Fit linear model, additionally report weights, additional metrics
}
\examples{
data <- purrr::map(mtcars, std) |> dplyr::bind_cols()
data$mpg <- mtcars$mpg

fit_lm(c("cyl", "disp"), "mpg", data)

fit_lm(c("cyl", "disp"), "mpg", data, concise = FALSE)

vars <- setdiff(colnames(mtcars), "mpg")
fit_lm(vars, "mpg", data)
fit_lm(c("disp", "hp", "wt", "qsec", "am"), "mpg", data)
}
