% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_msmt.R
\name{make_adjr2}
\alias{make_adjr2}
\title{Make adjusted/unadjusted R squared}
\usage{
make_adjr2(preds, actual, k, model_has_intercept = TRUE)
}
\arguments{
\item{preds}{numeric vector}

\item{actual}{numeric vector}

\item{k}{number of predictors}

\item{model_has_intercept}{logical}
}
\value{
numeric
}
\description{
Make adjusted/unadjusted R squared
}
\examples{
fit <- lm(mpg ~ cyl + disp, mtcars)

summary(fit)
# make_r2(predict(fit, mtcars), mtcars$mpg)
make_adjr2(predict(fit, mtcars), mtcars$mpg, 2)

fit <- lm(mpg ~ 0 + cyl + disp, mtcars)

summary(fit)
# make_r2(predict(fit, mtcars), mtcars$mpg, model_has_intercept = FALSE)
make_adjr2(predict(fit, mtcars), mtcars$mpg, 2, model_has_intercept = FALSE)
}
