% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_msmt.R
\name{make_rmse}
\alias{make_rmse}
\title{Make rmse}
\usage{
make_rmse(preds, actual)
}
\arguments{
\item{preds}{numeric vector}

\item{actual}{numeric vector}
}
\value{
numeric
}
\description{
Make rmse
}
\examples{
fit <- lm(mpg ~ cyl + disp, mtcars)
summary(fit)
make_rmse(fit$fitted.value, mtcars$mpg)
sqrt(sum(fit$residuals^2) / length(fit$residuals))
}
