% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_std.R
\name{std_fit}
\alias{std_fit}
\title{Standardize a dataframe to get a parameters list}
\usage{
std_fit(data)
}
\arguments{
\item{data}{dataframe of numeric columns.}
}
\value{
list with 3 defined values.
\itemize{
\item variables: a character vector of column names.
\item means: a numeric vector of means.
\item stds: a numeric vector of standard deviations.
}
}
\description{
Standardize a dataframe to get a parameters list
}
\examples{
a <- std_fit(mtcars)
a$means
a$stds
}
