% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_std.R
\name{std_transform}
\alias{std_transform}
\title{Standardize a numeric vector, using a list of standardization parameters}
\usage{
std_transform(parameters, data)
}
\arguments{
\item{parameters}{list competible with std_fit return value.}

\item{data}{dataframe of numeric columns.}
}
\value{
dataframe.
}
\description{
The list is expected to be generated from the std_fit function.
}
\examples{
std_paras <- std_fit(mtcars[c("hp", "mpg", "vs")])

std_transform(std_paras, mtcars) |> summary()
std_transform(std_paras, rbind(mtcars, mtcars)) |> summary()
std_transform(std_paras, mtcars[c("vs", "mpg")]) |> summary()
std_transform(std_paras, mtcars[c("carb", "wt")]) |> summary()
}
